# Property value storage optimization