--- a/src/Avalonia.Base/AvaloniaObject.cs
+++ b/src/Avalonia.Base/AvaloniaObject.cs
@@ -25,6 +25,7 @@ namespace Avalonia
         private List<AvaloniaObject>? _inheritanceChildren;
         private ValueStore? _values;
         private bool _batchUpdate;
+        private bool _isInitializing;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="AvaloniaObject"/> class.
@@ -126,6 +127,9 @@ namespace Avalonia
                 {
                     _values = new ValueStore(this);
 
+                    _values.IsDuringInitialization = _isInitializing;
+                    _isInitializing = false;
+                    
                     if (_batchUpdate)
                         _values.BeginBatchUpdate();
                 }
@@ -492,6 +496,18 @@ namespace Avalonia
 
             _batchUpdate = false;
             _values?.EndBatchUpdate();
+        }
+        
+        internal void SetValueStoreIsInitializing(bool isInitializing)
+        {
+            if (_values is null)
+            {
+                _isInitializing = isInitializing;
+                
+                return;
+            }
+
+            _values.IsDuringInitialization = isInitializing;
         }
 
         /// <inheritdoc/>