--- a/src/Avalonia.Base/PropertyStore/RemoveSentinelValue.cs
+++ b/src/Avalonia.Base/PropertyStore/RemoveSentinelValue.cs
@@ -28,5 +28,13 @@ namespace Avalonia.PropertyStore
                 newValue.Cast<T>(),
                 Priority));
         }
+
+        public void BeginBatchUpdate()
+        {
+        }
+
+        public void EndBatchUpdate()
+        {
+        }
     }
 }