… src/Avalonia.Base/Properties/AssemblyInfo.cs
… tests/Avalonia.Benchmarks/Layout/ControlsBenchmark.cs