## What does the pull request do?
<!--- Give a bit of background on the PR here, together with links to with related issues etc. -->
Based on my profiling in many cases controls got bottlenecked due to value storage allocating constantly. This PR is basically implementing more ideas from WPF to optimize value storage concepts.

Benchmark results are below. I've implemented a simple dictionary value storage to compare against and while lookup perf was better there were regressions in insert/removal in a few cases. This will need more time for investigation.
In benchmarks:
`_Old` - old value storage implementation
`_Dictionary` - dictionary value storage implementation

Add/Remove:
|                                  Method | PropertyCount | UseShuffledProperties | isInitializing |        Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Allocated |
|---------------------------------------- |-------------- |---------------------- |--------------- |------------:|----------:|----------:|-------:|-------:|----------:|
|                                AddValue |             2 |                 False |          False |    47.57 ns |  0.724 ns |  0.642 ns | 0.0421 |      - |      88 B |
|                       AddAndRemoveValue |             2 |                 False |          False |    42.04 ns |  0.078 ns |  0.061 ns | 0.0421 |      - |      88 B |
|            AddAndRemoveValueInterleaved |             2 |                 False |          False |    34.18 ns |  0.085 ns |  0.075 ns | 0.0421 |      - |      88 B |
|                                AddValue |             2 |                 False |           True |    46.60 ns |  0.227 ns |  0.212 ns | 0.0421 |      - |      88 B |
|                       AddAndRemoveValue |             2 |                 False |           True |    42.12 ns |  0.156 ns |  0.138 ns | 0.0421 |      - |      88 B |
|            AddAndRemoveValueInterleaved |             2 |                 False |           True |    34.20 ns |  0.075 ns |  0.066 ns | 0.0421 |      - |      88 B |
|                            AddValue_Old |             2 |                 False |              ? |    89.00 ns |  0.086 ns |  0.080 ns | 0.0726 |      - |     152 B |
|                   AddAndRemoveValue_Old |             2 |                 False |              ? |   124.89 ns |  0.551 ns |  0.515 ns | 0.0994 |      - |     208 B |
|        AddAndRemoveValueInterleaved_Old |             2 |                 False |              ? |   101.94 ns |  0.207 ns |  0.173 ns | 0.0650 |      - |     136 B |
|                     AddValue_Dictionary |             2 |                 False |              ? |    67.00 ns |  0.357 ns |  0.334 ns | 0.1032 |      - |     216 B |
|            AddAndRemoveValue_Dictionary |             2 |                 False |              ? |    74.96 ns |  0.171 ns |  0.160 ns | 0.1032 |      - |     216 B |
| AddAndRemoveValueInterleaved_Dictionary |             2 |                 False |              ? |    74.37 ns |  0.291 ns |  0.272 ns | 0.1032 |      - |     216 B |
|                                AddValue |             2 |                  True |          False |    47.01 ns |  0.068 ns |  0.061 ns | 0.0421 |      - |      88 B |
|                       AddAndRemoveValue |             2 |                  True |          False |    42.71 ns |  0.064 ns |  0.057 ns | 0.0421 |      - |      88 B |
|            AddAndRemoveValueInterleaved |             2 |                  True |          False |    34.41 ns |  0.115 ns |  0.107 ns | 0.0421 |      - |      88 B |
|                                AddValue |             2 |                  True |           True |    46.25 ns |  0.330 ns |  0.293 ns | 0.0421 |      - |      88 B |
|                       AddAndRemoveValue |             2 |                  True |           True |    43.51 ns |  0.117 ns |  0.109 ns | 0.0420 | 0.0001 |      88 B |
|            AddAndRemoveValueInterleaved |             2 |                  True |           True |    34.48 ns |  0.086 ns |  0.071 ns | 0.0421 |      - |      88 B |
|                            AddValue_Old |             2 |                  True |              ? |    86.35 ns |  0.159 ns |  0.148 ns | 0.0726 |      - |     152 B |
|                   AddAndRemoveValue_Old |             2 |                  True |              ? |   129.26 ns |  0.532 ns |  0.498 ns | 0.0994 |      - |     208 B |
|        AddAndRemoveValueInterleaved_Old |             2 |                  True |              ? |   101.90 ns |  0.282 ns |  0.235 ns | 0.0650 |      - |     136 B |
|                     AddValue_Dictionary |             2 |                  True |              ? |    68.37 ns |  0.168 ns |  0.158 ns | 0.1032 |      - |     216 B |
|            AddAndRemoveValue_Dictionary |             2 |                  True |              ? |    76.97 ns |  0.292 ns |  0.259 ns | 0.1032 |      - |     216 B |
| AddAndRemoveValueInterleaved_Dictionary |             2 |                  True |              ? |    77.26 ns |  0.164 ns |  0.153 ns | 0.1032 | 0.0001 |     216 B |
|                                AddValue |             6 |                 False |          False |   169.23 ns |  0.205 ns |  0.192 ns | 0.1490 |      - |     312 B |
|                       AddAndRemoveValue |             6 |                 False |          False |   172.72 ns |  0.382 ns |  0.339 ns | 0.1490 |      - |     312 B |
|            AddAndRemoveValueInterleaved |             6 |                 False |          False |    83.54 ns |  0.208 ns |  0.174 ns | 0.0421 |      - |      88 B |
|                                AddValue |             6 |                 False |           True |   144.15 ns |  0.158 ns |  0.123 ns | 0.1147 |      - |     240 B |
|                       AddAndRemoveValue |             6 |                 False |           True |   148.44 ns |  0.311 ns |  0.275 ns | 0.1147 |      - |     240 B |
|            AddAndRemoveValueInterleaved |             6 |                 False |           True |    83.73 ns |  0.195 ns |  0.183 ns | 0.0421 |      - |      88 B |
|                            AddValue_Old |             6 |                 False |              ? |   254.97 ns |  0.798 ns |  0.746 ns | 0.2866 | 0.0005 |     600 B |
|                   AddAndRemoveValue_Old |             6 |                 False |              ? |   429.13 ns |  0.604 ns |  0.565 ns | 0.4973 |      - |   1,040 B |
|        AddAndRemoveValueInterleaved_Old |             6 |                 False |              ? |   282.78 ns |  0.653 ns |  0.611 ns | 0.1721 |      - |     360 B |
|                     AddValue_Dictionary |             6 |                 False |              ? |   153.33 ns |  0.432 ns |  0.337 ns | 0.2217 |      - |     464 B |
|            AddAndRemoveValue_Dictionary |             6 |                 False |              ? |   182.72 ns |  0.210 ns |  0.186 ns | 0.2217 |      - |     464 B |
| AddAndRemoveValueInterleaved_Dictionary |             6 |                 False |              ? |   122.88 ns |  0.578 ns |  0.483 ns | 0.1032 |      - |     216 B |
|                                AddValue |             6 |                  True |          False |   179.86 ns |  0.235 ns |  0.184 ns | 0.1490 |      - |     312 B |
|                       AddAndRemoveValue |             6 |                  True |          False |   196.32 ns |  0.317 ns |  0.264 ns | 0.1490 |      - |     312 B |
|            AddAndRemoveValueInterleaved |             6 |                  True |          False |    84.29 ns |  0.110 ns |  0.097 ns | 0.0421 |      - |      88 B |
|                                AddValue |             6 |                  True |           True |   157.22 ns |  0.428 ns |  0.379 ns | 0.1147 |      - |     240 B |
|                       AddAndRemoveValue |             6 |                  True |           True |   174.32 ns |  0.283 ns |  0.251 ns | 0.1147 |      - |     240 B |
|            AddAndRemoveValueInterleaved |             6 |                  True |           True |    84.82 ns |  0.063 ns |  0.049 ns | 0.0421 |      - |      88 B |
|                            AddValue_Old |             6 |                  True |              ? |   232.02 ns |  0.294 ns |  0.245 ns | 0.2868 |      - |     600 B |
|                   AddAndRemoveValue_Old |             6 |                  True |              ? |   440.67 ns |  7.999 ns |  6.680 ns | 0.4973 |      - |   1,040 B |
|        AddAndRemoveValueInterleaved_Old |             6 |                  True |              ? |   283.24 ns |  0.377 ns |  0.315 ns | 0.1721 |      - |     360 B |
|                     AddValue_Dictionary |             6 |                  True |              ? |   156.87 ns |  0.737 ns |  0.616 ns | 0.2217 |      - |     464 B |
|            AddAndRemoveValue_Dictionary |             6 |                  True |              ? |   187.38 ns |  0.339 ns |  0.283 ns | 0.2217 |      - |     464 B |
| AddAndRemoveValueInterleaved_Dictionary |             6 |                  True |              ? |   121.86 ns |  0.685 ns |  0.640 ns | 0.1032 |      - |     216 B |
|                                AddValue |            10 |                 False |          False |   390.90 ns |  5.705 ns |  5.336 ns | 0.4549 |      - |     952 B |
|                       AddAndRemoveValue |            10 |                 False |          False |   379.77 ns |  0.453 ns |  0.424 ns | 0.4549 |      - |     952 B |
|            AddAndRemoveValueInterleaved |            10 |                 False |          False |   127.98 ns |  0.347 ns |  0.290 ns | 0.0420 |      - |      88 B |
|                                AddValue |            10 |                 False |           True |   254.45 ns |  0.407 ns |  0.381 ns | 0.2484 |      - |     520 B |
|                       AddAndRemoveValue |            10 |                 False |           True |   275.98 ns |  0.303 ns |  0.253 ns | 0.2484 |      - |     520 B |
|            AddAndRemoveValueInterleaved |            10 |                 False |           True |   128.07 ns |  0.252 ns |  0.236 ns | 0.0420 |      - |      88 B |
|                            AddValue_Old |            10 |                 False |              ? |   446.54 ns |  2.769 ns |  2.590 ns | 0.6232 |      - |   1,304 B |
|                   AddAndRemoveValue_Old |            10 |                 False |              ? |   854.32 ns |  1.826 ns |  1.618 ns | 1.1396 |      - |   2,384 B |
|        AddAndRemoveValueInterleaved_Old |            10 |                 False |              ? |   454.50 ns |  1.298 ns |  1.214 ns | 0.2789 |      - |     584 B |
|                     AddValue_Dictionary |            10 |                 False |              ? |   289.67 ns |  0.696 ns |  0.617 ns | 0.4740 |      - |     992 B |
|            AddAndRemoveValue_Dictionary |            10 |                 False |              ? |   336.25 ns |  0.605 ns |  0.566 ns | 0.4740 |      - |     992 B |
| AddAndRemoveValueInterleaved_Dictionary |            10 |                 False |              ? |   165.40 ns |  0.288 ns |  0.269 ns | 0.1032 |      - |     216 B |
|                                AddValue |            10 |                  True |          False |   426.06 ns |  4.860 ns |  4.308 ns | 0.4549 |      - |     952 B |
|                       AddAndRemoveValue |            10 |                  True |          False |   439.81 ns |  1.920 ns |  1.603 ns | 0.4549 |      - |     952 B |
|            AddAndRemoveValueInterleaved |            10 |                  True |          False |   132.81 ns |  0.953 ns |  0.795 ns | 0.0420 |      - |      88 B |
|                                AddValue |            10 |                  True |           True |   293.38 ns |  2.424 ns |  2.268 ns | 0.2484 |      - |     520 B |
|                       AddAndRemoveValue |            10 |                  True |           True |   340.90 ns |  0.797 ns |  0.706 ns | 0.2484 |      - |     520 B |
|            AddAndRemoveValueInterleaved |            10 |                  True |           True |   130.89 ns |  0.252 ns |  0.236 ns | 0.0420 |      - |      88 B |
|                            AddValue_Old |            10 |                  True |              ? |   429.99 ns |  1.575 ns |  1.315 ns | 0.6232 |      - |   1,304 B |
|                   AddAndRemoveValue_Old |            10 |                  True |              ? |   828.16 ns |  1.170 ns |  0.977 ns | 1.1396 |      - |   2,384 B |
|        AddAndRemoveValueInterleaved_Old |            10 |                  True |              ? |   459.29 ns |  1.109 ns |  1.037 ns | 0.2789 |      - |     584 B |
|                     AddValue_Dictionary |            10 |                  True |              ? |   295.23 ns |  1.147 ns |  1.073 ns | 0.4740 |      - |     992 B |
|            AddAndRemoveValue_Dictionary |            10 |                  True |              ? |   341.28 ns |  0.761 ns |  0.675 ns | 0.4740 |      - |     992 B |
| AddAndRemoveValueInterleaved_Dictionary |            10 |                  True |              ? |   164.78 ns |  0.473 ns |  0.420 ns | 0.1032 |      - |     216 B |
|                                AddValue |            20 |                 False |          False |   861.99 ns |  4.438 ns |  4.152 ns | 1.1473 |      - |   2,400 B |
|                       AddAndRemoveValue |            20 |                 False |          False |   839.32 ns |  1.789 ns |  1.673 ns | 1.1473 |      - |   2,400 B |
|            AddAndRemoveValueInterleaved |            20 |                 False |          False |   241.95 ns |  1.281 ns |  1.198 ns | 0.0420 |      - |      88 B |
|                                AddValue |            20 |                 False |           True |   486.84 ns |  0.609 ns |  0.570 ns | 0.5045 |      - |   1,056 B |
|                       AddAndRemoveValue |            20 |                 False |           True |   587.03 ns |  1.474 ns |  1.379 ns | 0.5045 |      - |   1,056 B |
|            AddAndRemoveValueInterleaved |            20 |                 False |           True |   241.73 ns |  1.252 ns |  1.110 ns | 0.0420 |      - |      88 B |
|                            AddValue_Old |            20 |                 False |              ? | 1,117.02 ns |  3.577 ns |  3.171 ns | 2.0008 |      - |   4,184 B |
|                   AddAndRemoveValue_Old |            20 |                 False |              ? | 2,393.25 ns |  5.062 ns |  4.488 ns | 3.8147 |      - |   7,984 B |
|        AddAndRemoveValueInterleaved_Old |            20 |                 False |              ? |   891.84 ns |  2.185 ns |  1.825 ns | 0.5465 |      - |   1,144 B |
|                     AddValue_Dictionary |            20 |                 False |              ? |   547.92 ns |  1.405 ns |  1.314 ns | 0.9937 |      - |   2,080 B |
|            AddAndRemoveValue_Dictionary |            20 |                 False |              ? |   648.28 ns |  1.150 ns |  1.076 ns | 0.9937 |      - |   2,080 B |
| AddAndRemoveValueInterleaved_Dictionary |            20 |                 False |              ? |   272.88 ns |  0.980 ns |  0.916 ns | 0.1030 |      - |     216 B |
|                                AddValue |            20 |                  True |          False |   933.97 ns |  3.766 ns |  3.523 ns | 1.1473 |      - |   2,400 B |
|                       AddAndRemoveValue |            20 |                  True |          False | 1,021.37 ns |  3.043 ns |  2.846 ns | 1.1463 |      - |   2,400 B |
|            AddAndRemoveValueInterleaved |            20 |                  True |          False |   247.70 ns |  1.072 ns |  0.950 ns | 0.0420 |      - |      88 B |
|                                AddValue |            20 |                  True |           True |   592.38 ns |  1.496 ns |  1.327 ns | 0.5045 |      - |   1,056 B |
|                       AddAndRemoveValue |            20 |                  True |           True |   790.40 ns |  2.496 ns |  2.335 ns | 0.5045 |      - |   1,056 B |
|            AddAndRemoveValueInterleaved |            20 |                  True |           True |   246.78 ns |  0.623 ns |  0.520 ns | 0.0420 |      - |      88 B |
|                            AddValue_Old |            20 |                  True |              ? | 1,038.93 ns |  2.952 ns |  2.465 ns | 2.0008 |      - |   4,184 B |
|                   AddAndRemoveValue_Old |            20 |                  True |              ? | 2,286.53 ns |  7.583 ns |  7.093 ns | 3.8147 |      - |   7,984 B |
|        AddAndRemoveValueInterleaved_Old |            20 |                  True |              ? |   899.11 ns |  0.574 ns |  0.479 ns | 0.5465 |      - |   1,144 B |
|                     AddValue_Dictionary |            20 |                  True |              ? |   543.19 ns |  1.178 ns |  1.102 ns | 0.9937 |      - |   2,080 B |
|            AddAndRemoveValue_Dictionary |            20 |                  True |              ? |   645.14 ns |  1.973 ns |  1.846 ns | 0.9937 |      - |   2,080 B |
| AddAndRemoveValueInterleaved_Dictionary |            20 |                  True |              ? |   270.10 ns |  0.605 ns |  0.537 ns | 0.1030 |      - |     216 B |
|                                AddValue |            30 |                 False |          False | 1,201.31 ns |  3.063 ns |  2.865 ns | 1.6060 |      - |   3,360 B |
|                       AddAndRemoveValue |            30 |                 False |          False | 1,234.66 ns |  3.479 ns |  3.255 ns | 1.6060 |      - |   3,360 B |
|            AddAndRemoveValueInterleaved |            30 |                 False |          False |   356.71 ns |  1.000 ns |  0.835 ns | 0.0420 |      - |      88 B |
|                                AddValue |            30 |                 False |           True |   642.80 ns |  0.845 ns |  0.660 ns | 0.5045 |      - |   1,056 B |
|                       AddAndRemoveValue |            30 |                 False |           True |   847.77 ns |  1.104 ns |  0.922 ns | 0.5045 |      - |   1,056 B |
|            AddAndRemoveValueInterleaved |            30 |                 False |           True |   352.78 ns |  1.182 ns |  0.987 ns | 0.0420 |      - |      88 B |
|                            AddValue_Old |            30 |                 False |              ? | 2,015.63 ns |  7.313 ns |  6.841 ns | 4.1428 |      - |   8,664 B |
|                   AddAndRemoveValue_Old |            30 |                 False |              ? | 4,517.84 ns | 10.278 ns |  9.111 ns | 8.0261 |      - |  16,784 B |
|        AddAndRemoveValueInterleaved_Old |            30 |                 False |              ? | 1,327.81 ns |  1.041 ns |  0.869 ns | 0.8144 |      - |   1,704 B |
|                     AddValue_Dictionary |            30 |                 False |              ? |   609.82 ns |  1.102 ns |  0.977 ns | 0.9937 |      - |   2,080 B |
|            AddAndRemoveValue_Dictionary |            30 |                 False |              ? |   757.22 ns |  1.479 ns |  1.235 ns | 0.9937 |      - |   2,080 B |
| AddAndRemoveValueInterleaved_Dictionary |            30 |                 False |              ? |   373.87 ns |  1.007 ns |  0.893 ns | 0.1030 |      - |     216 B |
|                                AddValue |            30 |                  True |          False | 1,334.35 ns | 14.653 ns | 13.707 ns | 1.6060 |      - |   3,360 B |
|                       AddAndRemoveValue |            30 |                  True |          False | 1,542.07 ns |  1.917 ns |  1.793 ns | 1.6060 |      - |   3,360 B |
|            AddAndRemoveValueInterleaved |            30 |                  True |          False |   362.03 ns |  1.363 ns |  1.275 ns | 0.0420 |      - |      88 B |
…