@grokys 2022-07-03
I'm working with this class right now and it's a pain to use. If it simply followed the `Dictionary` API, it would be a much better experience - is there any real reason that it can't follow the `Dictionary` API? (It doesn't have to implement `IDictionary<,>` just be familiar).